Update edd-software-license-manager.php
While it might not be the perfect solution, having a "global" License Term setting will be sufficient for most installations.  

This is my very first change on GitHub and I'm very open to learning.  Please let me know if there are things I missed.

Also note, there are changes in two other files that I'm working on next.